Extract feed module and add test
