+++ Cargo.toml
@@ -23,4 +23,5 @@ simple_logger = "1.13"
 tokio = { version = "1.9", features = ["macros"] }
 
 [dev-dependencies]
+pretty_assertions = "0.7"
 wiremock = "0.5"
