+++ src/bin/gen-feed/testdata/feed.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<rss version="2.0">
+  <channel>
+    <title>Dilbert</title>
+    <link>https://dilbert.com</link>
+    <description>Dilbert Daily Strip</description>
+    <item>
+      <title>Dilbert - 2018-10-01</title>
+      <link>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-10-01.gif</link>
+      <description>
+        <![CDATA[<img src="https://dilbert-feed-test.s3.amazonaws.com/strips/2018-10-01.gif">]]>
+      </description>
+      <guid>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-10-01.gif</guid>
+      <pubDate>Mon, 01 Oct 2018 00:00:00 +0000</pubDate>
+    </item>
+    <item>
+      <title>Dilbert - 2018-09-30</title>
+      <link>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-30.gif</link>
+      <description>
+        <![CDATA[<img src="https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-30.gif">]]>
+      </description>
+      <guid>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-30.gif</guid>
+      <pubDate>Sun, 30 Sep 2018 00:00:00 +0000</pubDate>
+    </item>
+    <item>
+      <title>Dilbert - 2018-09-29</title>
+      <link>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-29.gif</link>
+      <description>
+        <![CDATA[<img src="https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-29.gif">]]>
+      </description>
+      <guid>https://dilbert-feed-test.s3.amazonaws.com/strips/2018-09-29.gif</guid>
+      <pubDate>Sat, 29 Sep 2018 00:00:00 +0000</pubDate>
+    </item>
+  </channel>
+</rss>
